requirements.txt
Flask>=3.0
mediapipe>=0.10
numpy>=1.24
opencv-python>=4.8
pyserial>=3.5
tomli>=2.0.1
